Allow lexer input iterators to be Forward or better
This change is implemented primarily by templating the Scanner class
on the lexer IteratorT and using that type for three scanner data
members.  In addition we make use of std::distance/std::advance where
possible so we don't lose performance on RandomAccess inputs.
