+++ include/boost/wave/cpplexer/re2clex/strict_cpp.re
